LC-pascals-triangle-ii [Runtime: N/A; Memory: N/A]

+++ array/pascals-triangle-ii.cpp
@@ -14,7 +14,7 @@ public:
         
         int n = rowIndex;
         vector<int> res = {1};
-        int ans = 1;
+        long long ans = 1;
         for(int i=1; i<=rowIndex; i++){
             ans = (ans * n) / i;
             n--;
